web_app/flask_app/requirements.txt
cryptography
Flask==2.1.0
uWSGI==2.0.19.1
mysql-connector-python==8.0.21
